Implemented verification system
