+++ data/users.csv
@@ -1,4 +1,4 @@
 disc,steam,mmr,pos1,pos2,pos3,pos4,pos5
-215529156031545344,28707060,2500,1,1,1,5,5
+215529156031545344,28707060,2500,1,1,3,4,3
 238402271849742336,159179884,3000,5,5,5,5,5
 181839327985139713,279050958,4100,5,5,5,5,5
